Got a lot of things building; still WIP though

+++ unpack.sh
@@ -18,11 +18,17 @@ then
     exit 1
 fi
 
+if [ ! `which git` ]
+then
+    echo "Couldn't find git program, install git"
+    exit 1
+fi
+
 rm -rf source || exit 1
 mkdir source || exit 1
 cd source
 
-for archive in `cat ../archive_list`
+for archive in `cat ../archives.list`
 do
     extension=`echo "$archive" | awk -F . '{print $NF}'`
     echo "Unpacking $archive"
@@ -34,6 +40,10 @@ do
     fi
 done
 
-echo "Unpacking jsmn (cloning from Mercurial repository)"
-hg clone https://bitbucket.org/zserge/jsmn jsmn
+for hg_repo in `cat ../hg_repos.list`
+do
+    repo_name=`basename $hg_repo`
+    echo "Unpacking $repo_name (cloning from Mercurial repository)"
+    hg clone $hg_repo $repo_name
+done
